tests: move login/logout tests into separate test file
